rename macvlan_dhcp.rs to dhcp.rs
Given that bridge also uses the code it makes no sense to call it
macvlan. Also move some code around to not keep the dhcp code in
core_utils.
