+++ src/network/mod.rs
@@ -13,9 +13,9 @@ use crate::{
 pub mod bridge;
 pub mod constants;
 pub mod core_utils;
+mod dhcp;
 pub mod driver;
 pub mod internal_types;
-mod macvlan_dhcp;
 pub mod netlink;
 pub mod plugin;
 pub mod vlan;
